Patch 0281
・ipに関する型をunsigned long型に統一
・portに関する型をunsigned short型に統一
・ファイルへの書き込みを fopen() ではなく lock_fopen() を使うように修正（core.c, db.c）
・lock_fopen() で10000回トライしても安全なファイル名が得られない場合はNULLを返すように修正（lock.c）
・gccでの警告潰し（skill.c）

+++ src/char/char.h
@@ -8,8 +8,8 @@
 #define DEFAULT_AUTOSAVE_INTERVAL_CS 300*1000
 
 struct mmo_map_server {
-	long ip;
-	short port;
+	unsigned long ip;
+	unsigned short port;
 	int users;
 	char *map;
 	short map_num;
@@ -42,8 +42,8 @@ extern int autosave_interval;
 struct char_online {
 	int account_id;
 	int char_id;
-	int ip;
-	int port;
+	unsigned long ip;
+	unsigned short port;
 	char name[24];
 };
