+++ src/map/chrif.h
@@ -4,7 +4,7 @@
 void chrif_setuserid(char*);
 void chrif_setpasswd(char*);
 void chrif_setip(char*);
-void chrif_setport(int);
+void chrif_setport(unsigned short);
 int chrif_parse(int fd);
 
 int chrif_isconnect(void);
@@ -13,7 +13,7 @@ int chrif_authreq(struct map_session_data *);
 int chrif_save(struct map_session_data*);
 int chrif_charselectreq(struct map_session_data *);
 
-int chrif_changemapserver(struct map_session_data *sd,char *name,int x,int y,int ip,short port);
+int chrif_changemapserver(struct map_session_data *sd,char *name,int x,int y,unsigned long ip,unsigned short port);
 
 void chrif_searchcharid(int char_id);
 int chrif_changegm(int id,const char *pass,int len);
